requirement.txt
langchain_google_genai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
